adding first frees and changed my ultoa

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: sad-aude <sad-aude@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/24 15:07:14 by sad-aude          #+#    #+#             */
-/*   Updated: 2020/05/23 01:24:53 by sad-aude         ###   ########lyon.fr   */
+/*   Updated: 2020/05/25 17:51:38 by sad-aude         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ char			*ft_strjoin(char *s1, char *s2, int id);
 char			*ft_strtrim(char const *s1, char const *set);
 char			**ft_split(char const *s, char c);
 char			*ft_itoa(int n);
-char			*ft_ultoa(unsigned int n);
+char			*ft_ultoa(unsigned long int n);
 char			*ft_itoa_base(unsigned int value, int base);
 char			*ft_ultoa_base(unsigned long long value, int base);
 char			*ft_strmapi(char const *s, char (*f)(unsigned int, char));
